# Improved logging performance by refactoring Logging class.