--- a/src/classes/Logging.ts
+++ b/src/classes/Logging.ts
@@ -1,44 +1,106 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
+/* eslint-disable no-console */
 import { ILogger, ILogger_ } from 'src/interfaces/ILogger';
 
 /**
- * Logging class; encapsulates console.log, console.debug, console.warn and console.error
+ * Enum representing logging levels as numeric values.
+ */
+export enum LoggingLevelNumber {
+    'none' = 0,
+    'trace' = 1,
+    'debug' = 2,
+    'info' = 3,
+    'warn' = 4,
+    'error' = 5,
+}
+
+/**
+ * Representing logging levels as string values.
+ */
+export type LoggingLevel =
+    | 'none'
+    | 'trace'
+    | 'debug'
+    | 'info'
+    | 'warn'
+    | 'error';
+
+/**
+ * Logging class; encapsulates console.log, console.debug, console.warn and console.error.
  */
 export const Logging: ILogger_ = class Logging implements ILogger {
     private static _instance: Logging;
-    private _logLevel: LoggingLevel;
+    private _logLevel: LoggingLevelNumber;
     private _logPrefix: string;
+    /**
+     * Gets the default log level.
+     */
+    private static get _defaultLogLevel(): LoggingLevelNumber {
+        return LoggingLevelNumber.info;
+    }
+
+    public trace: (...args: unknown[]) => void;
+    public debug: (...args: unknown[]) => void;
+    public info: (...args: unknown[]) => void;
+    public warn: (...args: unknown[]) => void;
+    public error: (...args: unknown[]) => void;
 
     /**
-     * Creates a new Logging instance
-     * @param logLevel The log level to use. Defaults to "info"
-     * @param logPrefix The prefix to prepend to log messages
+     * Creates a new Logging instance.
+     * @param logLevel The log level to use. Defaults to "info".
+     * @param logPrefix The prefix to prepend to log messages.
      */
-    constructor(logLevel: LoggingLevel = 'info', logPrefix = '') {
-        this._logLevel = logLevel;
+    constructor(logLevel?: LoggingLevel, logPrefix = '') {
         this._logPrefix = logPrefix ? `${logPrefix}-` : '';
+        this.setLogLevel(logLevel);
 
-        if (this._logLevel === 'none') {
-            // eslint-disable-next-line no-console
+        if (this._logLevel === LoggingLevelNumber.none) {
             console.info('Logging disabled');
         }
         Logging._instance = this;
     }
 
     /**
-     * Sets the log level
-     * @param logLevel The log level to set
+     * Sets the log level and assigns the appropriate logging methods.
+     * @param logLevel The log level to set.
      */
-    public setLogLevel(logLevel: LoggingLevel) {
-        this._logLevel = logLevel;
-        // eslint-disable-next-line no-console
-        console.info(`Log level set to ${logLevel}`);
+    public setLogLevel(logLevel: LoggingLevel | undefined) {
+        this._setLogLevel(this.parseLoggingLevel(logLevel));
     }
 
     /**
-     * Returns the Logging instance
-     * @returns The Logging instance
+     * Sets the log level and assigns the appropriate logging methods.
+     * @param logLevel The log level to set.
+     */
+    private _setLogLevel(logLevel: LoggingLevelNumber) {
+        this._logLevel = logLevel;
+
+        this.trace = this._none;
+        this.debug = this._none;
+        this.info = this._none;
+        this.warn = this._none;
+        this.error = this._none;
+
+        if (this.logLevelActive(LoggingLevelNumber.trace))
+            this.trace = this._trace.bind(this);
+
+        if (this.logLevelActive(LoggingLevelNumber.debug))
+            this.debug = this._debug.bind(this);
+
+        if (this.logLevelActive(LoggingLevelNumber.info))
+            this.info = this._info.bind(this);
+
+        if (this.logLevelActive(LoggingLevelNumber.warn))
+            this.warn = this._warn.bind(this);
+
+        if (this.logLevelActive(LoggingLevelNumber.error))
+            this.error = this._error.bind(this);
+
+        console.info(`Log level set to ${LoggingLevelNumber[this._logLevel]}`);
+    }
+
+    /**
+     * Returns the Logging instance.
+     * @returns The Logging instance.
      */
     public static getInstance(): Logging {
         if (!Logging._instance) {
@@ -49,9 +111,9 @@ export const Logging: ILogger_ = class Logging implements ILogger {
     }
 
     /**
-     * Returns an object with logging methods that prepend a specified prefix to messages
-     * @param prefix The prefix to prepend to all log messages
-     * @returns An object with logging methods
+     * Returns an object with logging methods that prepend a specified prefix to messages.
+     * @param prefix The prefix to prepend to all log messages.
+     * @returns An object with logging methods.
      */
     public static getLogger(prefix: string): ILogger {
         const instance = Logging.getInstance();
@@ -67,7 +129,7 @@ export const Logging: ILogger_ = class Logging implements ILogger {
 
     /**
      * Creates a new logger with a specified prefix.
-     * @param instance The Logging instance which will be used to log messages. Eg. Singleton instance.
+     * @param instance The Logging instance which will be used to log messages.
      * @param prefix The prefix to prepend to the log message.
      * @returns An object with logging methods.
      */
@@ -81,37 +143,34 @@ export const Logging: ILogger_ = class Logging implements ILogger {
              * @param args The arguments to log
              * @returns Nothing
              */
-            trace: (...args: any[]): void =>
-                instance.logWithPrefix('trace', prefix, args),
+            trace: (...args: unknown[]): void =>
+                instance.trace(prefix, ...args),
             /**
              * Logs a `debug` message
              * @param args The arguments to log
              * @returns Nothing
              */
-            debug: (...args: any[]): void =>
-                instance.logWithPrefix('debug', prefix, args),
+            debug: (...args: unknown[]): void =>
+                instance.debug(prefix, ...args),
             /**
              * Logs an `info` message
              * @param args The arguments to log
              * @returns Nothing
              */
-            info: (...args: any[]): void =>
-                instance.logWithPrefix('info', prefix, args),
+            info: (...args: unknown[]): void => instance.info(prefix, ...args),
             /**
              * Logs a `warn` message
              * @param args The arguments to log
              * @returns Nothing
              */
-            warn: (...args: any[]): void =>
-                instance.logWithPrefix('warn', prefix, args),
+            warn: (...args: unknown[]): void => instance.warn(prefix, ...args),
             /**
              * Logs an `error` message
              * @param args The arguments to log
              * @returns Nothing
              */
-            error: (...args: any[]): void =>
-                instance.logWithPrefix('error', prefix, args),
-
+            error: (...args: unknown[]): void =>
+                instance.error(prefix, ...args),
             /**
              * Sets the log level
              * @param logLevel The log level to set
@@ -124,141 +183,105 @@ export const Logging: ILogger_ = class Logging implements ILogger {
     }
 
     /**
-     * Logs a message with a specified prefix and level.
-     * @param level - The logging level.
-     * @param prefix - The prefix to add to the log message.
-     * @param args - The arguments to be logged.
+     * A no-operation function used for disabled log levels.
+     * @param args The arguments to ignore.
      */
-    private logWithPrefix(
-        level: Exclude<LoggingLevel, 'none'>,
-        prefix: string,
-        args: any[],
-    ) {
-        args.unshift(prefix);
-        (this[level] as (...args: any[]) => void)(...args);
+    private _none(...args: unknown[]): void {
+        // No operation
     }
 
     /**
-     * Logs a message to the console if the log level is "trace"
-     * @param message The message to log
-     * @param optionalParams Optional parameters to log
+     * Logs a message to the console if the log level is "trace".
+     * @param message The message to log.
+     * @param optionalParams Optional parameters to log.
      */
-    public trace(message?: any, ...optionalParams: any[]): void {
-        if (this.logLevelActive('trace')) {
-            const logMessage = this.constructLogMessage(message);
-            // eslint-disable-next-line no-console
-            console.debug(logMessage, ...optionalParams);
-        }
+    private _trace(message?: unknown, ...optionalParams: unknown[]): void {
+        const logMessage = this.constructLogMessage(message);
+        console.trace(logMessage, ...optionalParams);
     }
 
     /**
-     * Logs a message to the console if the log level is "debug"
-     * @param message The message to log
-     * @param optionalParams Optional parameters to log
+     * Logs a message to the console if the log level is "debug".
+     * @param message The message to log.
+     * @param optionalParams Optional parameters to log.
      */
-    public debug(message?: any, ...optionalParams: any[]): void {
-        if (this.logLevelActive('debug')) {
-            const logMessage = this.constructLogMessage(message);
-            // eslint-disable-next-line no-console
-            console.debug(logMessage, ...optionalParams);
-        }
+    private _debug(message?: unknown, ...optionalParams: unknown[]): void {
+        const logMessage = this.constructLogMessage(message);
+        console.debug(logMessage, ...optionalParams);
     }
 
     /**
-     * Logs a message to the console if the log level is "info" or "debug"
-     * @param message The message to log
-     * @param optionalParams Optional parameters to log
+     * Logs a message to the console if the log level is "info".
+     * @param message The message to log.
+     * @param optionalParams Optional parameters to log.
      */
-    public info(message?: any, ...optionalParams: any[]): void {
-        if (this.logLevelActive('info')) {
-            const logMessage = this.constructLogMessage(message);
-            // eslint-disable-next-line no-console
-            console.info(logMessage, ...optionalParams);
-        }
+    private _info(message?: unknown, ...optionalParams: unknown[]): void {
+        const logMessage = this.constructLogMessage(message);
+        console.info(logMessage, ...optionalParams);
     }
 
     /**
-     * Logs a message to the console if the log level is "info", "debug" or "warn"
-     * @param message The message to log
-     * @param optionalParams Optional parameters to log
+     * Logs a message to the console if the log level is "warn".
+     * @param message The message to log.
+     * @param optionalParams Optional parameters to log.
      */
-    public warn(message?: any, ...optionalParams: any[]): void {
-        if (this.logLevelActive('warn')) {
-            const logMessage = this.constructLogMessage(message);
-            // eslint-disable-next-line no-console
-            console.warn(logMessage, ...optionalParams);
-        }
+    private _warn(message?: unknown, ...optionalParams: unknown[]): void {
+        const logMessage = this.constructLogMessage(message);
+        console.warn(logMessage, ...optionalParams);
     }
 
     /**
-     * Logs a message to the console if the log level is "info", "debug", "warn" or "error"
-     * @param message The message to log
-     * @param optionalParams Optional parameters to log
+     * Logs a message to the console if the log level is "error".
+     * @param message The message to log.
+     * @param optionalParams Optional parameters to log.
      */
-    public error(message?: any, ...optionalParams: any[]): void {
-        if (this.logLevelActive('error')) {
-            const logMessage = this.constructLogMessage(message);
-            // eslint-disable-next-line no-console
-            console.error(logMessage, ...optionalParams);
-        }
+    private _error(message?: unknown, ...optionalParams: unknown[]): void {
+        const logMessage = this.constructLogMessage(message);
+        console.error(logMessage, ...optionalParams);
     }
 
     /**
-     * Constructs a log message with the log prefix
-     * @param message The message to log
-     * @returns The constructed log message: `${this._logPrefix}${message}`
+     * Constructs a log message with the log prefix.
+     * @param message The message to log.
+     * @returns The constructed log message.
      */
-    private constructLogMessage(message?: any): string {
+    private constructLogMessage(message?: unknown): string {
         return `${this._logPrefix}${message}`;
     }
 
     /**
-     * Checks if the log level is active
-     * @param logLevel The log level to check
-     * @returns Whether the log level is active
+     * Checks if the log level is active.
+     * @param logLevel The log level to check.
+     * @returns Whether the log level is active.
      */
-    private logLevelActive(logLevel: LoggingLevel): boolean {
-        if (this._logLevel === 'none') {
-            return false;
-        }
+    private logLevelActive(logLevel: LoggingLevelNumber): boolean {
+        return (
+            this._logLevel !== LoggingLevelNumber.none &&
+            this._logLevel <= logLevel
+        );
+    }
 
-        if (this._logLevel === 'trace') {
-            return true;
+    /**
+     * Parses a logging level string to a numeric value.
+     * @param level The logging level string.
+     * @returns The numeric logging level.
+     */
+    private parseLoggingLevel(level: string | undefined): LoggingLevelNumber {
+        switch (level) {
+            case 'none':
+                return LoggingLevelNumber.none;
+            case 'trace':
+                return LoggingLevelNumber.trace;
+            case 'debug':
+                return LoggingLevelNumber.debug;
+            case 'info':
+                return LoggingLevelNumber.info;
+            case 'warn':
+                return LoggingLevelNumber.warn;
+            case 'error':
+                return LoggingLevelNumber.error;
+            default:
+                return Logging._defaultLogLevel;
         }
-
-        if (this._logLevel === 'debug') {
-            return logLevel !== 'trace';
-        }
-
-        if (this._logLevel === 'info') {
-            return logLevel !== 'trace' && logLevel !== 'debug';
-        }
-
-        if (this._logLevel === 'warn') {
-            return (
-                logLevel !== 'trace' &&
-                logLevel !== 'debug' &&
-                logLevel !== 'info'
-            );
-        }
-
-        if (this._logLevel === 'error') {
-            return (
-                logLevel !== 'trace' &&
-                logLevel !== 'debug' &&
-                logLevel !== 'info' &&
-                logLevel !== 'warn'
-            );
-        }
-
-        return true;
     }
 };
-
-export type LoggingLevel =
-    | 'none'
-    | 'trace'
-    | 'debug'
-    | 'info'
-    | 'warn'
-    | 'error';