- Introduced LoggingLevelNumber enum for efficient log level comparisons.
- Optimized logging method assignments to reduce runtime checks.
- Added no-operation function for disabled log levels.
- Bound logging methods directly to improve performance.

Performance improvements:

| Category         | Old Avg. Time (ms) | New Avg. Time (ms) |
|------------------|---------------------|---------------------|
| logging disabled | 0.8613              | 0.2621              |
| logging enabled  | 14.1476             | 9.6143              |

This refactor significantly reduces the time required for 1000 log calls, both with logging enabled and disabled.